GameMessage and VsGameSection header fixes
